✨ 메인 페이지 및 레이아웃 개선
- Open Graph 이미지 URL을 '/api/og'로 변경하여 동적 이미지 생성 적용
- 캠페인 데이터를 서버에서 직접 가져오는 방식으로 변경하여 초기 로딩 성능 개선
- 새로운 Header 및 CampaignGrid 컴포넌트 추가로 코드 구조화 및 재사용성 향상
- 메타데이터 동적 생성 기능 추가로 SEO 최적화 강화
